探索結果取得処理を微修正
