+++ frontend/hooks.js
@@ -1,5 +1,4 @@
 import axios from 'axios'
-import { assoc } from 'ramda'
 import useSWR from "swr"
 import { toCamelCaseObject } from './utils';
 
@@ -12,16 +11,27 @@ export const useGetSpotList = () => {
 
   const { data, error, mutate } = useSWR(API_URL, fetcher)
 
-  const spotList = parseSpotList(data)
   return {
-    spotList: spotList,
+    spotList: data ? toCamelCaseObject(data) : data,
     error: error,
     mutate: mutate
   }
 }
 
-const parseSpotList = (data) => {
-  if (data) {
-    return toCamelCaseObject(data)
+
+export const useGetSearchResult = (param) => {
+  const API_URL = process.env.NEXT_PUBLIC_API_ROOT + '/search'
+  const fetcher = async (url, param) => {
+    const body = JSON.parse(decodeURI(param))
+    const res = await axios.post(url, body)
+    return res.data
+  }
+
+  const { data, error, mutate } = useSWR(param ? API_URL : null, (url) => fetcher(url, param))
+
+  return {
+    searchResult: data ? toCamelCaseObject(data) : data,
+    error: error,
+    mutate: mutate
   }
 }
